requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
binance==0.3
certifi==2024.7.4
charset-normalizer==3.3.2
dateparser==1.2.0
DateTime==5.5
eventkit==1.0.3
frozenlist==1.4.1
ib-insync==0.9.86
idna==3.7
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.0.1
pandas==2.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
tomli==2.0.1
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
zope.interface==7.0.1
